refactor(automation): extract job card
